refactor(sqlite): split adapter.go into reads/writes/graph/embeddings; decompose ReadSymbolGraph
Split the 1444-line adapter.go along its concern clusters: lifecycle/
schema/tx/reset stays in adapter.go; flat reads, writes, graph traversal,
and embedding queries move to reads.go, writes.go, graph.go, embeddings.go.
All within one package, so the moves are behavior-preserving.

Decompose ReadSymbolGraph (gocyclo 28) into foldRootMembers and an
expandGraphHop method hung off a small graphHop state struct, leaving
ReadSymbolGraph a readable hop loop at gocyclo <=15. Also retire the
formatSymbolContext 27-07 ledger entry by extracting relationshipGroups
and fitGroups.

Prepared-statement lifecycle on the hot-path write methods is preserved.
Conformance suite green; maket graph query byte-identical on the existing
index; sqlite package coverage 91.1% -> 91.6%.
